Mysql reconnect option was broken
There were two problems discovered after the release of 72.0

 1. during config autoReconnect was firing in common cases and because
@connection_parameters was not yet set the adapter was broken.
 2. Processing of the reconnect: flag in db config would enable
reconnect whether it was set to true or false.

The other point is none of our tests ran into this because we do not
permute through all the possible config values.  Future work should
be to at least create a simple app from scratch as part of a CI
job so at least do not have failure out of the gate like this.